moved few tip-related methods to its own module; added more details into documentation
