+++ Dockerfile
@@ -25,7 +25,7 @@ COPY babel.config.json ./
 COPY tsconfig.json ./
 COPY src/ ./src
 
-RUN yarn install --frozen-lockfile
+RUN yarn install --immutable
 
 RUN yarn build
